Successfully finsihed functionality for 'show attendees' button
